requirements.txt
streamlit
reportlab
pandas
Pillow
openpyxl